remove duplicate audio processing from the app and modify import to take src one
